Fix #14, #31 - Allow disabling HTTP and configuration option for HTTPS address
Introduces a new command line option --http_disable to completely
disable the HTTP connector. It is not done automatically if HTTPS
is enabled to preserve backward compatibility. Instead, a warning
will be logged if HTTPS is used but HTTP is still enabled.

If neither HTTP or HTTPS are enabled, an exception is thrown.

This patch also allows specifying a HTTPS listening address
as is already possible for HTTP.