Bump to Guice 4.1.0
We use Jersey 1.x; The last build for Jersey 1.x targets Guice 3.0,
which is broken for Java 8 (throws an exception while... wait for it...
rendering the exception message). There's magic byte code manipulation
happening here so we can't simply just bump Guice to 4.1.0 and expect
things to work, Guice 4.1.0 is source compat with Guice 3.0, but not bin
compat. Yay!

Rather than upgrade to Jersey 2 (likely lots of work), we include a
special build for Jersey-Guice 1.x with Guice 4.1.0.